feat(native)!: remove TimeField, superseded by MaskField
A time typed into a box is `MaskField` with `mask="time"`, and the two never shared
an engine: `MaskField` masks through `utils/mask.ts`, `TimeField` through a
`time-mask.ts` of its own. Two implementations of the same typing, one of which had
nothing the other lacked.

What went with it: the component, its subpath and tsup entry, its demo screen and
route, and the half of the time util that only it called — `maskTime`, `parseTime`,
`formatTime`, `timeLength`, `timePlaceholder` and their tests. What is left of that
file is `hourCycleFor`, `periodOf` and `withTime`, which `TimePicker` and
`DateTimePicker` read, so it is now `utils/time.ts`: nothing in it masks anything.

`TimePicker` — the dial — is untouched. The roadmap keeps P5.26h on the list as
`dropped`, so the decision stays visible.
