+++ apps/demo/app/index.tsx
@@ -77,7 +77,6 @@ const SCREENS = [
   { href: '/text-area', label: 'TextArea' },
   { href: '/text-field', label: 'TextField' },
   { href: '/phone-number-field', label: 'PhoneNumberField' },
-  { href: '/time-field', label: 'TimeField' },
   { href: '/timeline', label: 'Timeline' },
   { href: '/time-picker', label: 'TimePicker' },
   { href: '/toggle-button', label: 'ToggleButton' },
