+++ apps/docs/lib/data/native-api.generated.json
@@ -8719,82 +8719,6 @@
       ]
     }
   ],
-  "time-field": [
-    {
-      "name": "TimeFieldProps",
-      "props": [
-        {
-          "name": "value",
-          "type": "Date | null | undefined",
-          "required": false,
-          "description": "The time shown. Present means controlled — the field follows it, and `onValueChange` is how it asks for a new one.  A `Date`, like everything else in this library that carries a moment: a time on its own still lands on a day, and keeping the type lets a `DateField`'s value pass straight through. The day is the one already held, or today's when there is none."
-        },
-        {
-          "name": "defaultValue",
-          "type": "Date | null | undefined",
-          "required": false,
-          "description": "Where an uncontrolled field starts."
-        },
-        {
-          "name": "onValueChange",
-          "type": "((value: Date | null) => void) | undefined",
-          "required": false,
-          "description": "Every edit, with the time the field now holds — or `null` while it is not one yet."
-        },
-        {
-          "name": "granularity",
-          "type": "TimeGranularity | undefined",
-          "required": false,
-          "description": "How far down the time is written."
-        },
-        {
-          "name": "hourCycle",
-          "type": "HourCycle | undefined",
-          "required": false,
-          "description": "Twelve hours and a period beside them, or twenty-four. Unset, it is `locale`'s."
-        },
-        {
-          "name": "locale",
-          "type": "string | undefined",
-          "required": false,
-          "description": "The hour cycle comes from here when it is not given."
-        },
-        {
-          "name": "timeLabels",
-          "type": "TimeLabels | undefined",
-          "required": false,
-          "description": "What the placeholder calls each part — `{ hours: 'HH', minutes: 'mm', seconds: 'ss' }`.  The letters are a language's and this library does not pick one, so the default is the one code is written in. A `placeholder` on the field wins over it."
-        },
-        {
-          "name": "periodLabels",
-          "type": "Record<DayPeriod, string> | undefined",
-          "required": false,
-          "description": "What `TimeField.Period` says."
-        },
-        {
-          "name": "children",
-          "type": "ReactNode",
-          "required": false,
-          "description": ""
-        }
-      ]
-    },
-    {
-      "name": "TimeFieldFieldProps",
-      "props": []
-    },
-    {
-      "name": "TimeFieldPeriodProps",
-      "props": [
-        {
-          "name": "accessibilityLabel",
-          "type": "string | undefined",
-          "required": false,
-          "description": "Overrides the text that's read by the screen reader when the user interacts with the element. By default, the label is constructed by traversing all the children and accumulating all the Text nodes separated by space. What a screen reader says. There is no default, for the reason the arrows have none."
-        }
-      ]
-    }
-  ],
   "time-picker": [
     {
       "name": "TimePickerProps",
