+++ packages/native/src/components/date-time-picker/date-time-picker.tsx
@@ -4,7 +4,7 @@ import type { TextStyle } from 'react-native'
 import { useControllableState } from '../../hooks/use-controllable-state'
 import { useXAUITheme } from '../../theme/theme-hooks'
 import { startOfDay } from '../../utils/dates'
-import { withTime } from '../../utils/time-mask'
+import { withTime } from '../../utils/time'
 import { selectRecipe } from '../select/select.recipe'
 import { DateTimePickerProvider } from './date-time-picker.context'
 import type { DateTimePickerProps } from './date-time-picker.type'
