+++ packages/native/src/components/index.ts
@@ -59,7 +59,6 @@ export * from './table'
 export * from './tabs'
 export * from './tag-group'
 export * from './text-area'
-export * from './time-field'
 export * from './time-picker'
 export * from './timeline'
 export * from './toggle-button'
