+++ packages/native/src/components/time-picker/time-picker.type.ts
@@ -12,7 +12,7 @@ import type { IconProps } from '../../system/icon'
 import type { PressableFeedbackProps } from '../../system/pressable-feedback'
 import type { TextStyleProps, ViewStyleProps } from '../../system/style-props'
 import type { RadiusKey, Size } from '../../theme/theme.type'
-import type { DayPeriod, HourCycle } from '../../utils/time-mask'
+import type { DayPeriod, HourCycle } from '../../utils/time'
 
 export type TimePickerSlot =
   | 'dial'
